targetRegionCount should be int, and rich UT

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/normalizer/SimpleRegionNormalizer.java
@@ -282,7 +282,7 @@ class SimpleRegionNormalizer implements RegionNormalizer, ConfigurationObserver 
     }
     TableName table = tableDescriptor.getTableName();
     double avgRegionSize;
-    long targetRegionCount = tableDescriptor.getNormalizerTargetRegionCount();
+    int targetRegionCount = tableDescriptor.getNormalizerTargetRegionCount();
     long targetRegionSize = tableDescriptor.getNormalizerTargetRegionSize();
     LOG.debug("Table {} configured with target region count {}, target region size {}", table,
       targetRegionCount, targetRegionSize);
